requirements.txt
langchain>=0.1.0
langchain-experimental>=0.0.43
langchain-community>=0.0.13
langchain-core>=0.1.10
langgraph>=0.0.20
pandas>=2.0.0
matplotlib>=3.5.0
openpyxl>=3.1.2
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.8.0
langchain-anthropic>=0.0.5